removes the template.ts pass-through barrel
Each module under src/gen/ now imports directly from the file that
defines what it needs (schema-types.ts, utils.ts, render.ts) rather
than going through the template.js re-export. This eliminates the
inconsistency where most files imported through the barrel while
render.ts itself imported siblings directly, and removes a hop in
IDE go-to-definition.

The combined template.test.ts is split into utils.test.ts and
render.test.ts so each test file sits next to the implementation it
covers. Generated output is byte-identical pre/post refactor.
