--- src/gen/template.ts
@@ -1,9 +1,0 @@
-// Barrel re-export — the actual implementations live in focused modules:
-//   schema-types.ts — type definitions (HerokuSchema, SchemaNode, etc.)
-//   utils.ts         — pure string/schema utilities
-//   render.ts        — TypeRenderer class
-
-export type { HerokuSchema, ResourceDefinition, SchemaNode, SchemaLink, HRefParam, RouteDefinition, HttpMethod } from './schema-types.js'
-export { HTTP_METHODS } from './schema-types.js'
-export { toPascalCase, toCamelCase, formatPropertyKey, renderJSDoc, disambiguateLinkTitles } from './utils.js'
-export { TypeRenderer } from './render.js'
